refactor: 클래스 파일 분리,불필요한 메서드 정리 tetrominoQueue.js appendQueue메서드 제거, Tetris/Main.js Init메서드 제거
